Reorganize code over multiple files
