+++ test/runtests.jl
@@ -4,6 +4,6 @@ using Test
 @testset "HerbConstraints.jl" verbose=true begin
     include("test_solver.jl")
     include("test_pattern_match.jl")
-    include("test_propagators.jl")
+    #include("test_propagators.jl")
 end
